% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_addClassificationDirectives.R
\name{addClassificationDirectives}
\alias{addClassificationDirectives}
\title{Classification directives}
\usage{
addClassificationDirectives(x, method = NULL, methodParameters = NULL, ...)
}
\arguments{
\item{x}{IMC_study, a study}
}
\description{
Classification directives
}
\examples{
\dontrun{

}
}
