% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_addInterpretationMatrix.R
\name{addInterpretationMatrix}
\alias{addInterpretationMatrix}
\title{Interpretation Matrix}
\usage{
addInterpretationMatrix(
  x,
  classifierObject = NULL,
  undtLabel = NULL,
  undtID = NULL,
  ...
)
}
\arguments{
\item{x}{IMC_study, a study}
}
\description{
Interpretation Matrix
}
\examples{
\dontrun{

}
}
