% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_addSegmentationDirectives.R
\name{addSegmentationDirectives}
\alias{addSegmentationDirectives}
\title{Segmentation directives}
\usage{
addSegmentationDirectives(x, method = NULL, methodParameters = NULL, ...)
}
\arguments{
\item{x}{IMC_study, a study}
}
\description{
Segmentation directives
}
\examples{
\dontrun{

}
}
