% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_IMC_Analysis.R
\name{an_listAnalysis}
\alias{an_listAnalysis}
\title{List all the analysis}
\usage{
an_listAnalysis(x, ...)
}
\arguments{
\item{x}{IMC_Study, a study.}
}
\value{
character vector.
}
\description{
List the analysis that have been created and not dismissed
for the specified study.
}
\examples{
\dontrun{
listAnalysis(MyStudy)
}
}
