% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_distilExpression.R
\name{distilExpression}
\alias{distilExpression}
\title{Distil expression}
\usage{
distilExpression(x, method = "weightedMean", ...)
}
\arguments{
\item{x}{IMC_study, a study}
}
\description{
Distil expression
}
\examples{
\dontrun{

}
}
