% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_exportTiffImage.R
\name{exportTiffImage}
\alias{exportTiffImage}
\title{Export Tiff image}
\usage{
exportTiffImage(
  x,
  what = c("raw", "derived", "classification"),
  layerNames = NULL,
  NAVal = 0,
  uids = NULL,
  fileFolder = NULL,
  transposeImage = F,
  flipImage = c("NA", "x", "y", "xy"),
  scaleImage = F,
  bitsImage = c("32", "16", "8"),
  verbose = T,
  ...
)
}
\arguments{
\item{x}{environment, a study}

\item{fileFolder}{root folder containing at least a folder (which name is gone to be used as layer name), containing one tiff mask for each sample}

\item{...}{not implemented}
}
\value{
none
}
\description{
Export Tiff image
}
\examples{

\dontrun{
}

}
