% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDef_IMC_Analysis.R
\docType{class}
\name{IMC_Analysis-class}
\alias{IMC_Analysis-class}
\alias{IMC_Analysis}
\title{IMC_Analysis}
\description{
Analysis class is actually an environment containing various objects used
within a study
}
