% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDef_IMC_AnalysisList.R
\docType{class}
\name{IMC_AnalysisList-class}
\alias{IMC_AnalysisList-class}
\alias{IMC_AnalysisList}
\title{IMC_AnalysisList}
\description{
Simple list containing names of analysis used within one study.
}
