% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDef_IMC_ChannelTable.R
\docType{class}
\name{IMC_ChannelTable-class}
\alias{IMC_ChannelTable-class}
\alias{IMC_ChannelTable}
\title{IMC_ChannelTable}
\description{
Data frame containing information regarding channel and markers names
}
