% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDef_IMC_Classification.R
\docType{class}
\name{IMC_Classification-class}
\alias{IMC_Classification-class}
\alias{IMC_Classification}
\title{IMC_Classification}
\description{
List-like object containing all the sample IMC_RasterStacks for classification
purpose
}
