% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDef_IMC_ClassificationDirectives.R
\docType{class}
\name{IMC_ClassificationDirectives-class}
\alias{IMC_ClassificationDirectives-class}
\alias{IMC_ClassificationDirectives}
\title{IMC_ClassificationDirectives}
\description{
Object containing the classification parameters to be applied. Each
classification method needs to be implemented as an engine function, with
a list of default parameter
and have a proper section of the method segment() able to process the
method parametes and call the appropriate engine.
}
\section{Slots}{

\describe{
\item{\code{method}}{character, method name, needs to mach the call in the
witch statement in the segment method.}

\item{\code{methodParameters}}{list, method parameters as specified for the
method}
}}

