% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDef_IMC_ClassificationDirectivesList.R
\docType{class}
\name{IMC_ClassificationDirectivesList-class}
\alias{IMC_ClassificationDirectivesList-class}
\alias{IMC_ClassificationDirectivesList}
\title{IMC_ClassificationDirectivesList}
\description{
Container list for IMC_ClassificationDirectives
}
