% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classDef_IMC_Study.R
\docType{class}
\name{IMC_Study-class}
\alias{IMC_Study-class}
\alias{IMC_Study}
\title{Study}
\description{
A study is an environment containing all the objects that the different
functions and methods need to work with. In brief a study must contain
all the sample images organized in raster stacks, a study table specifying
various meta-data concerning the samples, and a further Analysis environment
containing the elaboration of the raw data. Usually a study is created via
the function \link{initStudy}. The only reason for initializing an empty study it would
be to populate the environment with elements from other studies via the
\link{archive}/\link{retrieve} system.
Note that the show method for a study utilizes mainly the information
contained in the study table. If the study table is empty, i.e. the object
myStudy$studyTable is NULL, the show method will return 'empty study' even
if all the other objects within the study are properly set and functional.
}
