% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initObjectAttr.R
\name{initObjectAttr}
\alias{initObjectAttr}
\title{Initialize object attributes}
\usage{
initObjectAttr(x)
}
\arguments{
\item{x}{an object}
}
\value{
the same object with attributes creation (crtnTimeStmp),
modification (mdtnTimeStmp), archivation (artnTimeStmp), and
the path to the archiviation folder (fileArchive)
}
\description{
Internal utility function to initialize object attributes.
}
