% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_loadTxtToStack.R
\name{loadTxtToStack}
\alias{loadTxtToStack}
\title{From txt to rasterStack}
\usage{
loadTxtToStack(
  fn_path = NULL,
  fn_file = NULL,
  fn_rasterDBPath = NULL,
  fn_rasterStackPath = NULL,
  fn_cols = NULL,
  fn_newNames = NULL,
  fn_transpose = F,
  fn_norm = F,
  fn_x_column = "X",
  fn_y_column = "Y",
  fn_details = list(study = "NO_study", sample = "NO_sample", ROI = "NO_ROI", replicate =
    "NO_replicate", bioGroup = "No_bioGroup"),
  fn_channel = NULL,
  fn_trsh = 0.965,
  fn_zeroOff = T,
  fn_nThread = 1,
  fn_verbose = F
)
}
\arguments{
\item{fn_path}{Root path for a bunch of .txt image files.}

\item{fn_file}{.txt file name containing raw data.}

\item{fn_rasterDBPath}{target folder where to store raster object.}

\item{fn_rasterStackPath}{target folder where to store rasterStack files.}

\item{fn_cols}{vector of column names to load.}

\item{fn_newNames}{vector of nemes to use as layer names.}

\item{fn_norm}{normalise raw data using quantile after stripping all 0-values.}

\item{fn_details}{list of experimental details, possible names are study = study name, sample, ROI, replicate, bioGroup.
These information can be updated via \code{\link[=updateMetadata]{updateMetadata()}}.}

\item{fn_channel}{IMC_ChannelTable object.}

\item{fn_trsh}{quantile as the high treshold used for normalisation.}

\item{fn_zeroOff}{logical whether or not to strip 0-value in the process of normalisation.}

\item{fn_nThread}{numeric scalar, number of thread to use in \link[data.table:fread]{data.table::fread},
if data.table version is <1.14.3 fn_nThread is coerced to 1}
}
\value{
an IMC_RasterStack object
}
\description{
\code{\link[=loadTxtToStack]{loadTxtToStack()}} is an internal function, used by \code{\link[=initStudy]{initStudy()}} to convert a txt file to a RasterStack.
}
