% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_makeClassificationModel.R
\name{makeClassificationModel}
\alias{makeClassificationModel}
\title{Make a classification model}
\usage{
makeClassificationModel(x, method = NULL, seed = 1234, ...)
}
\arguments{
\item{x}{IMC_study, a study}
}
\description{
Make a classification model
}
\examples{
\dontrun{

}
}
