% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_mdfObjectAttr.R
\name{mdfObjectAttr}
\alias{mdfObjectAttr}
\title{Modify object attributes}
\usage{
mdfObjectAttr(x, crtnTimeStmp, mdtnTimeStmp, artnTimeStmp, fileArchive, sealed)
}
\arguments{
\item{x}{an object}

\item{crtnTimeStmp}{character as produced by format(Sys.time(),format="\%F \%T \%Z", tz = Sys.timezone())
representing the creation time.}

\item{mdtnTimeStmp}{character, modification time}

\item{artnTimeStmp}{character, archiviation time}

\item{fileArchive}{character, archiviation path}
}
\value{
the same object with attributes creation (crtnTimeStmp),
modification (mdtnTimeStmp), archivation (artnTimeStmp), and
the path to the archiviation folder (fileArchive)
}
\description{
Internal utility function to modify object attributes.
}
