% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_mergeIMC_stacks.R
\name{mergeIMC_stacks}
\alias{mergeIMC_stacks}
\title{Merge IMC stacks}
\usage{
mergeIMC_stacks(x, y, update = T)
}
\arguments{
\item{x}{IMC_RasterStack}

\item{y}{IMC_RasterStack}

\item{update, }{logical, preserve layers with identical name from
both stacks (F)}
}
\description{
Utility function (should not be used directly) to merge two IMC stacks
obviously with reference to the same image (checked via uid).
Merge can preserve (update = F) or not all the layers in the case the two
stacks contain layers with the same name.
}
\details{
All the additional slots (uid, study...) are copied from x
while attributes are copied from y
}
