% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_retrieve.R
\name{retrieve}
\alias{retrieve}
\title{Retrieve}
\usage{
retrieve(x, ...)
}
\arguments{
\item{x}{character, file path to a file or directory}
}
\value{
the object written to the specified file or the raster stacks
contained in the specified folder
\dontrun{
myStudy <- retrieve(x = ./path/to/myStudy)
}
}
\description{
Retrieve load from disk a study, an analysis or part of it
}
\details{
archive has specific methods for a study and for its elements
}
