% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_runIMC.R
\name{runIMC}
\alias{runIMC}
\title{RUNIMC GUI for annotation}
\usage{
runIMC(x = NULL, help = F)
}
\description{
Open a GUI for features annotations, setting help=T give some hints on how to
proceed during annotation.... but it is mostly just annoying when you alrady
know what to do, so it is set to False by default.
}
