% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_strgfObjectAttr.R
\name{strgfObjectAttr}
\alias{strgfObjectAttr}
\title{Stringify object attributes}
\usage{
strgfObjectAttr(x)
}
\arguments{
\item{x}{an object}
}
\value{
string
}
\description{
Internal utility function to imake a string of object attributes.
}
