% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_testSegment.R
\name{testSegment}
\alias{testSegment}
\title{Test segmentation}
\usage{
testSegment(
  x,
  labelLayer = "label",
  label = NULL,
  uid = NULL,
  limits = NULL,
  ...
)
}
\arguments{
\item{x}{IMC_study, a study}
}
\description{
Test segmentation
}
\examples{
\dontrun{

}
}
