% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mthd_IMC_TrainingFeatures.R
\name{tf_parLabelList}
\alias{tf_parLabelList}
\title{Training features}
\usage{
tf_parLabelList(x, ...)
}
\arguments{
\item{x}{environment, IMC_TrainingFeatures, a study or the object storing training features}
}
\value{
numeric vector
}
\description{
Lists labels present in the training features table
}
