% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_tifToList.R
\name{tifToList}
\alias{tifToList}
\title{tifToList}
\usage{
tifToList(
  fn_sourceFile = NULL,
  fn_origin = c(0, 0),
  fn_direction = c(1, 1),
  fn_resolution = c(1, 1),
  fn_Zindex = 0,
  fn_columnNames = NULL,
  fn_targetFile = NULL,
  fn_nThread = 1,
  fn_verbose = T
)
}
\arguments{
\item{fn_sourceFile}{character vector containing the full path to each tif file
as it can be obtained for eamle with \link{list.files} with full.names=T}

\item{fn_origin}{numeric vector of length two indicating the coordinate of the
origin of the image, by default the origin correspond to the upper left corner
of the image, therefore the image will be flipped vertically.}

\item{fn_direction}{numeric vector of length two, of which only the sign will
used to determine if the picture extends left (-1) or right (+1),
and up (+1) or down (-1).}

\item{fn_resolution}{numeric vector of length two setting x and y resolution.}

\item{fn_Zindex}{numeric scalar, starting number for cell indexing (Z).}

\item{fn_targetFile}{character scalar, full path to the output file.}

\item{fn_nThread}{numeric scalar, number of thread to use in \link[data.table:fread]{data.table::fread},
if data.table version is <1.14.3 fn_nThread is coerced to 1}

\item{fn_verbose}{logic scalar}

\item{fn_colNames}{character vector of the same lent as fn_sourceFile, giving
the name of each column. Column names should be compatible with the way they
are suppose to be interpred by \link{loadTxtToStack} and \link{initStudy} (eg: CD4-c01,
CD8-c02, DAPI-c03).}
}
\description{
Utility function to transform single tif images in a stack text file readable
with \link{loadTxtToStack} and \link{initStudy}. All tif images must have same extent.
}
