% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_validateData.R
\name{validateData}
\alias{validateData}
\title{Validate Data}
\usage{
validateData(
  fn_rawDataFolder = NULL,
  fn_pattern = ".txt",
  fn_x_column = "X",
  fn_y_column = "Y",
  fn_nonChannelColumns = c("X", "Y", "Z", "Start_push", "End_push", "Pushes_duration"),
  fn_emptyChannelPattern = "[0-9]+[A-z]+\\\\([A-z]+[0-9]+[A-z]+\\\\)",
  fn_usedChannelPattern = "[0-9]+[A-z]+-.+\\\\([A-z]+[0-9]+[A-z]+\\\\)",
  fn_channelPattern = "[0-9]+[A-z]+.*\\\\(([A-z]+[0-9]+[A-z]+)\\\\)",
  fn_markerPattern = "[0-9]+[A-z]+-(.+)\\\\([A-z]+[0-9]+[A-z]+\\\\)",
  ...
)
}
\value{
if everything is OK, the function returns 0, otherwise some clue to start
investigating what and where the problem might be.
}
\description{
Validate data checks that all the *.txt files used as raw input images have
the same header line, that channel names are recognized, and images are complete
meaning that all columns within an image has the same number and position of
rows and vice versa.
Parameters are the same as in \strong{\link{initStudy}}
}
